refractor: moved the sidemenu-toggle dervice file to the providers folder and update imports
