+++ libs/features/convs-mgr/stories/blocks/library/main/src/lib/components/block/block.component.ts
@@ -38,7 +38,7 @@ import { _CreateVideoInputBlockForm } from '../../model/video-input-block-form.m
 import { _CreateKeywordJumpBlockMessageForm } from '../../model/keyword-jump-form.model';
 
 import { BlockInjectorService } from '../../providers/block-injector.service';
-import { SidemenuToggleService } from 'libs/services/sidemenu-toggle.service';
+import { SidemenuToggleService } from 'libs/elements/layout/page-convl/src/lib/providers/sidemenu-toggle.service';
 /**
  * Block which sends a message from bot to user.
  */
